Html update
The header has the logo image and the navigation bar and the nav as an id of firstnav. Added the links for the style sheet and the reset css sheet. Added the charset and the title. The section has three sections inside of the section with the id of images. The first section element has an id of images1 and there are three figures that have the first three images in the figures. Same for the next two sections. Except The second and third sections are images2 and image3. Added the aside element with the page content. There is a primary heading and then paragraph tags.