Final Css commit
Added the font-face style rule and it makes the text this font. This rule is for the background and it has a color of blanched almond. The rule is for the body and the the rule has the background image, font style, width, margins left and right, border right and left and the box shadow for the body. The rule is for the navigation bar and its display is grid has a padding bottom of hundred pixels. The rule is for the anchor tags that are in list items that is in the unordered list that is in the nav with the id of firstnav. The rule is for the figures with the ids of image1,image2,image3,image4,image5,image6,image7,image8,image9, The rule has border and the box shadows. The rule is for the primary headings that are in the aside and have the color of the text and the text shadow. The rule is for the paragraph tags that are in the aside and has the color of the text and the text shadow.